+++ LibOS/shim/include/shim_signal.h
@@ -3,6 +3,7 @@
 
 #include <shim_types.h>
 #include <shim_defs.h>
+#include <ucontext.h>
 
 void sigaction_make_defaults(struct __kernel_sigaction* sig_action);
