+++ Pal/include/arch/x86_64/Linux-SGX/ucontext.h
@@ -1,0 +1,28 @@
+/*
+   This file is part of Graphene Library OS.
+
+   Graphene Library OS is free software: you can redistribute it and/or
+   modify it under the terms of the GNU Lesser General Public License
+   as published by the Free Software Foundation, either version 3 of the
+   License, or (at your option) any later version.
+
+   Graphene Library OS is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU Lesser General Public License for more details.
+
+   You should have received a copy of the GNU Lesser General Public License
+   along with this program.  If not, see <http://www.gnu.org/licenses/>.  */
+
+/*
+ * ucontext.h
+ *
+ * Linux-SGX PAL uses Linux's ucontext.h
+ */
+
+#ifndef _LINUX_SGX_UCONTEXT_H
+#define _LINUX_SGX_UCONTEXT_H
+
+#include <Linux/ucontext.h>
+
+#endif /* _LINUX_SGX_UCONTEXT_H */
